Fixed sound menus, Fixed buttons going grey, removed potato
The sound setting would change but not the actual sound instance. This was because we were only changing the setting variables but not refreshing the sound instance.
Added some love.graphics.setColor(255,255,255) lines around and about. Shoulda done that in the first place aye.
Eating an eggplant would say "you at a potato" or something along those lines. Fixed that.